chore: typescript enhancements
